Refactor Public Client
