+++ docs/public_blockchain_client.rst
@@ -1,0 +1,13 @@
+Public Blockchain Class - :mod:`public_blockchain_client`
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. contents:: Table of Contents
+
+
+.. automodule:: tradehub.public_blockchain_client
+    :autosummary:
+    :members:
+    :private-members:
+    :undoc-members:
+    :inherited-members:
+    :show-inheritance:
